removed initial default screen
