+++ lib/features/login/presentation/bloc/login_bloc.dart
@@ -7,12 +7,11 @@ part 'login_event.dart';
 part 'login_state.dart';
 
 class LoginBloc extends Bloc<LoginEvent, LoginState> {
+
+  LoginBloc() : super(LoginInitial());
+
   @override
-  LoginState get initialState => LoginInitial();
-  @override
-  Stream<LoginState> mapEventToState(
-    LoginEvent event,
-  ) async* {
+  Stream<LoginState> mapEventToState(LoginEvent event) async* {
     // TODO: implement mapEventToState
   }
 }
