+++ lib/utilities/constants.dart
@@ -1,0 +1,3 @@
+class Constants {
+  static const kAppTitle = "Swap n Swipe";
+}
